+++ RayTracer/include/interaction/interaction.h
@@ -12,9 +12,9 @@ class Interaction
 class SurfaceInteraction : public Interaction
 {
 public:
-	glm::vec3 p;
-	glm::vec3 normal;
-	glm::vec2 uv;
+	glm::dvec3 p;
+	glm::dvec3 normal;
+	glm::dvec2 uv;
 	std::shared_ptr<Material> mat;
 	std::shared_ptr<Texture> texture;
