+++ RayTracer/include/misc/progressreporter.h
@@ -56,14 +56,14 @@ class ProgressReporter {
         if (num == 0 /*|| PbrtOptions.quiet*/) return;
         workDone += num;
     }
-    float ElapsedMS() const {
+    double ElapsedMS() const {
         std::chrono::system_clock::time_point now =
             std::chrono::system_clock::now();
         int64_t elapsedMS =
             std::chrono::duration_cast<std::chrono::milliseconds>(now -
                                                                   startTime)
                 .count();
-        return (float)elapsedMS;
+        return (double)elapsedMS;
     }
     void Done();
